Testnet pallet is enabled in runtimes.
