+++ pallets/runtime/develop/Cargo.toml
@@ -41,7 +41,7 @@ pallet-treasury = { path = "../../treasury", default-features = false }
 pallet-utility = { path = "../../utility", default-features = false }
 polymesh-contracts = { path = "../../contracts", default-features = false }
 pallet-sudo = { path = "../../sudo", default-features = false }
-pallet-testnet = { path = "../../testnet", default-features = false }
+pallet-testnet = { path = "../../testnet", default-features = false, optional = true }
 
 # Others
 serde = { version = "1.0.104", default-features = false }
@@ -104,7 +104,7 @@ equalize = []
 testnet = [
     "polymesh-common-utilities/testnet",
     "pallet-staking/testnet",
-    "pallet-testnet/testnet",
+    "pallet-testnet",
     "polymesh-weights/testnet",
 ]
 no_std = [
@@ -160,7 +160,6 @@ std = [
     "pallet-transaction-payment/std",
     "pallet-treasury/std",
     "pallet-utility/std",
-    "pallet-testnet/std",
     "polymesh-common-utilities/std",
     "polymesh-primitives/std",
     "polymesh-runtime-common/std",
