+++ pallets/testnet/Cargo.toml
@@ -34,7 +34,6 @@ u32_backend = [ "confidential_identity/u32_backend"]
 u64_backend = [ "confidential_identity/u64_backend"]
 avx2_backend = [ "confidential_identity/avx2_backend"]
 
-testnet = []
 running-ci = []
 
 no_std = [
